Fix TCP hangs (unsigned wrap in tcpin.c trim()); LP64-correct int32/uint32; all -Wformat type fixes
TCP sessions stalled and deadlocked: the receiver ACKed exactly one
window (2048) beyond the sender's last byte and advertised window 0,
then both sides retransmitted forever while the VJ compressor (axhc.c)
dropped the duplicates. The root cause was the right-edge trim in
tcpin.c trim():

  if((excess = seg->seq + *length - (tcb->rcv.nxt + tcb->rcv.wnd)) > 0)

'seg->seq'/'rcv.nxt' are int32 (signed) and '*length'/'rcv.wnd' are
uint (unsigned), so the C usual arithmetic conversions promote the
whole expression to unsigned. A normal in-window segment (true excess
negative, e.g. -1933) therefore wrapped to a huge positive value,
'excess > 0' fired, and '*length -= excess' wrapped 115 to 2048.
rcv.nxt then advanced and rcv.wnd shrank by the full window on every
received data segment, freezing the session.

Fix: compute the delta in signed arithmetic via explicit (long) casts,
so in-window segments are left untouched and only real window
overruns are trimmed.

Also fix the int32/uint32 typedefs in global.h for LP64 hosts: there
'long' is 64 bits, which silently breaks the 32-bit wrap-around
arithmetic the TCP sequence numbers and the VJ compression deltas
depend on (axhc.c, tcpin.c, tcp.h, tcphdr.c, slhc.c, ...). Use 'int',
which is 32 bits on every platform this code targets. 'cmdparse.c'
setlong() accordingly takes int32*. The typedef change surfaced many
-Wformat specifier mismatches (%lu vs int32/uint32); all are fixed so
the tree builds with -Wformat -Wall without format warnings.

+++ src/ax25cmd.c
@@ -146,7 +146,7 @@ struct iface *ifp)
 		if(lp->iface != ifp)
 			continue;
 		printf("%-10s%-17s%8lu\n",pax25(tmp,lp->addr),
-		 tformat(secclock() - lp->time),lp->currxcnt);
+		 tformat(secclock() - lp->time),(unsigned long)lp->currxcnt);
 	}
 	return 0;
 }
@@ -205,7 +205,7 @@ struct iface *ifp)
 		} else {
 			printf("%-17s",tformat(secclock() - lq->time));
 		}
-		printf("%8lu\n",lp->currxcnt);
+		printf("%8lu\n",(unsigned long)lp->currxcnt);
 	}
 	return 0;
 }
@@ -281,7 +281,7 @@ void *p)
 			 axp->maxframe,
 			 axp->flags.remotebusy ? '*' : ' ',
 			 axp->retries,
-			 axp->srt,
+			 (unsigned long)axp->srt,
 			 Ax25states[axp->state],
 			 ax25hdr_to_string(&axp->hdr));
 		}
@@ -316,34 +316,35 @@ struct ax25_cb *axp)
 	printf(" %02u/%02u",axp->retries,axp->n2);
 	printf(" %s\n",Ax25states[axp->state]);
 
-	printf("srtt = %lu mdev = %lu ",axp->srt,axp->mdev);
+	printf("srtt = %lu mdev = %lu ",(unsigned long)axp->srt,
+	 (unsigned long)axp->mdev);
 	printf("T1: ");
 	if(run_timer(&axp->t1))
-		printf("%lu",read_timer(&axp->t1));
+		printf("%lu",(unsigned long)read_timer(&axp->t1));
 	else
 		printf("stop");
-	printf("/%lu ms; ",dur_timer(&axp->t1));
+	printf("/%lu ms; ",(unsigned long)dur_timer(&axp->t1));
 
 	printf("T3: ");
 	if(run_timer(&axp->t3))
-		printf("%lu",read_timer(&axp->t3));
+		printf("%lu",(unsigned long)read_timer(&axp->t3));
 	else
 		printf("stop");
-	printf("/%lu ms\n",dur_timer(&axp->t3));
+	printf("/%lu ms\n",(unsigned long)dur_timer(&axp->t3));
 
 	printf("T4: ");
 	if(run_timer(&axp->t4))
-		printf("%lu",read_timer(&axp->t4));
+		printf("%lu",(unsigned long)read_timer(&axp->t4));
 	else
 		printf("stop");
-	printf("/%lu ms; ",dur_timer(&axp->t4));
+	printf("/%lu ms; ",(unsigned long)dur_timer(&axp->t4));
 
 	printf("T5: ");
 	if(run_timer(&axp->t5))
-		printf("%lu",read_timer(&axp->t5));
+		printf("%lu",(unsigned long)read_timer(&axp->t5));
 	else
 		printf("stop");
-	printf("/%lu ms\n",dur_timer(&axp->t5));
+	printf("/%lu ms\n",(unsigned long)dur_timer(&axp->t5));
 #ifdef	AX25_VJCOMP
         /* MW: dump VJ statistics if any */
         if (axp->slcomp) {
